requirements.txt
# NLP & ML
spacy==3.7.2
nltk==3.8.1
sentence-transformers>=5.0.0
scikit-learn==1.3.0
torch>=2.0.0

# Document Processing
pdfminer.six==20221105
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.0.0

# Data Processing
pandas==2.0.3
numpy>=1.24.0

# FastAPI & Web
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1

# Security & Rate Limiting
python-jose==3.3.0
slowapi==0.1.9

# Monitoring & Utilities
psutil==5.9.7
python-json-logger==2.0.7